Authentication

+++ src/components/book/book.controller.js
@@ -9,6 +9,9 @@ const { errorResponse } = require("../../middleware/error-handling-middleware");
 
 const createBook = async (req, res) => {
   try {
+    if (!req.user.isAdmin) {
+      return res.status(403).send("Access Denied: Only admins can createBook");
+    }
     const book = await createBookService(req.body);
     res.status(201).send(book);
   } catch (error) {
@@ -17,6 +20,9 @@ const createBook = async (req, res) => {
 };
 
 const updateBook = async (req, res) => {
+  if (!req.user.isAdmin) {
+    return res.status(403).send("Access Denied: Only admins can updateBook");
+  }
   try {
     const book = await updateBookService(req.body);
     res.status(201).send(book);
@@ -26,6 +32,9 @@ const updateBook = async (req, res) => {
 };
 
 const deleteBook = async (req, res) => {
+  if (!req.user.isAdmin) {
+    return res.status(403).send("Access Denied: Only admins can deleteBook");
+  }
   try {
     await deleteBookService(req.query);
     res.status(200).send({ message: "Book deleted successfully" });
